+++ HISTORY.md
@@ -1,5 +1,7 @@
 # Release History
 
+## x.x.x (not yet released)
+
 ## 5.0.0 (2026-02-12)
 * Allow a Hint to be the first argument to Hint::build().
 * Make `Hint::*` constants into a `HintType` enumeration.