+++ HISTORY.md
@@ -1,6 +1,6 @@
 # Release History
 
-## x.x.x (not yet released)
+## 5.1.0 (2026-09-16)
 * Add JsonCodecableWithServicesTrait to make it easier to deserialize
   objects using the Dependency Injection pattern.
 * Add JsonCodecableWithServicesTestTrait, a PHPUnit test trait to verify